--- src/enhance_pedestrian_safety/cvips_data/success_20251215_092719/config.txt
@@ -1,2 +1,0 @@
-生成时间: 2025-12-15 09:27:19.432740
-地图: Carla/Maps/Town10HD_Opt
